# packages/api-contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScheduleFireRequest": {
        "additionalProperties": false,
        "properties": {
          "deviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviceid"
          },
          "scheduledFor": {
            "format": "date-time",
            "title": "Scheduledfor",
            "type": "string"
          }
        },
        "required": [
          "scheduledFor"
        ],
        "title": "ScheduleFireRequest",
        "type": "object"
      },
...
      "TaskScheduleCreate": {
        "additionalProperties": false,
        "properties": {
          "accountId": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Accountid",
            "type": "string"
          },
          "commandType": {
            "enum": [
              "xianyu.publish_listing.v1",
              "xianyu.collect_orders.v1",
              "xiaohongshu.publish_note.v1",
              "device.probe_capabilities.v1"
            ],
            "title": "Commandtype",
            "type": "string"
          },
          "deviceIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Deviceids",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "expectedBindingVersion": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedbindingversion"
          },
          "kind": {
            "enum": [
              "ONCE",
              "RECURRING"
            ],
            "title": "Kind",
            "type": "string"
          },
          "missPolicy": {
            "default": "QUEUE_ONE",
            "enum": [
              "QUEUE_ONE",
              "SKIP"
            ],
            "title": "Misspolicy",
            "type": "string"
          },
          "onceAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onceat"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "rrule": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rrule"
          },
          "startDeadlineMinutes": {
            "default": 30,
            "maximum": 10080.0,
            "minimum": 1.0,
            "title": "Startdeadlineminutes",
            "type": "integer"
          },
          "timezone": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "timezone",
          "kind",
          "accountId",
          "deviceIds",
          "commandType"
        ],
        "title": "TaskScheduleCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CloudCtl Control API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/task-schedules": {
      "get": {
        "operationId": "list_schedules_api_v1_task_schedules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Schedules Api V1 Task Schedules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Schedules",
        "tags": [
          "task-schedules"
        ]
      },
      "post": {
        "operationId": "create_schedule_api_v1_task_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskScheduleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Schedule Api V1 Task Schedules Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Schedule",
        "tags": [
          "task-schedules"
        ]
      }
    },
    "/api/v1/task-schedules/{schedule_id}": {
      "get": {
        "operationId": "get_schedule_api_v1_task_schedules__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Schedule Api V1 Task Schedules  Schedule Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schedule",
        "tags": [
          "task-schedules"
        ]
      }
    },
    "/api/v1/task-schedules/{schedule_id}:disable": {
      "post": {
        "operationId": "disable_schedule_api_v1_task_schedules__schedule_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Disable Schedule Api V1 Task Schedules  Schedule Id  Disable Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Schedule",
        "tags": [
          "task-schedules"
        ]
      }
    },
    "/api/v1/task-schedules/{schedule_id}:enable": {
      "post": {
        "operationId": "enable_schedule_api_v1_task_schedules__schedule_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enable Schedule Api V1 Task Schedules  Schedule Id  Enable Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Schedule",
        "tags": [
          "task-schedules"
        ]
      }
    },
    "/api/v1/task-schedules/{schedule_id}:fire": {
      "post": {
        "operationId": "fire_schedule_api_v1_task_schedules__schedule_id__fire_post",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleFireRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fire Schedule Api V1 Task Schedules  Schedule Id  Fire Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fire Schedule",
        "tags": [
          "task-schedules"
        ]
      }
    },
...
  }
}